Fix intern string pool ref-counting and graph deletion cleanup
- Defer pool operations (acquire/release) until MvccGraph::commit()
  instead of applying them inline in attribute store methods. This
  fixes the undo/rollback bug where intermediate commits (from WITH
  clauses) would adjust pool refs that couldn't be undone on rollback.
- Release all interned refs in MvccGraph::Drop so GRAPH.DELETE properly
  cleans up pool entries.
- Split out-of-scope test classes (persistence, replication) into
  separate files in flow_tests_todo.txt.
