+++ tests/flow/test_intern_string.py
@@ -372,139 +372,4 @@ class testInternString():
         res = self.graph.query(q).result_set
         self.env.assertEqual(['a', 'a', 'b', 'b', 'c', 'c', 1, 2, 3], res[0][0])
 
-class testInternStringPersistency():
-    def __init__(self):
-        self.env, self.db = Env(enableDebugCommand=True)
-        self.conn = self.env.getConnection()
-
-        # skip test if we're running under Sanitizer
-        if SANITIZER:
-            self.env.skip() # sanitizer is not working correctly with bulk
-
-        # Synchronous deletion
-        self.db.config_set('ASYNC_DELETE', 'no')
-
-        # clear DB
-        self.conn.flushall()
-
-    def tearDown(self):
-        # clear DB
-        self.conn.flushall()
-        self.graph = self.db.select_graph(GRAPH_ID)
-
-        assertStringPoolStats(self.conn, 0, 0)
-
-    def testInternStringPersistent(self):
-        # populate DB
-
-        # create first node
-        q = "CREATE ({value: intern($s)})"
-
-        # create multiple EMPTY graphs
-        graphs = []
-        for _ in range(0, 10):
-            g = self.db.select_graph(random_string())
-            g.query(q, {'s': SMALL_STRING})
-            graphs.append(g)
-
-        # validate string pool stats
-        assertStringPoolStats(self.conn, 1, 10)
-
-        # Save RDB & Load from RDB
-        self.env.dumpAndReload()
-
-        # string-pool stats expected to match former stats before reload
-        assertStringPoolStats(self.conn, 1, 10)
-
-        for g in graphs:
-            res = g.query("MATCH (n) RETURN n.value").result_set[0][0]
-            self.env.assertEqual(res, SMALL_STRING)
-
-class testInternStringReplication():
-    def __init__(self):
-        # skip test if we're running under sanitizer
-        if SANITIZER:
-            Environment.skip(None) # sanitizer is not working correctly with replication
-
-        self.env, self.db = Env(env='oss', useSlaves=True)
-        self.conn = self.env.getConnection()
-        self.graph = self.db.select_graph(GRAPH_ID)
-
-        self.source_con = self.env.getConnection()
-        self.replica_con = self.env.getSlaveConnection()
-
-        # force effects replication
-        self.db.config_set('EFFECTS_THRESHOLD', 0)
-
-        # Synchronous deletion
-        self.source_con.execute_command("GRAPH.CONFIG", "SET", 'ASYNC_DELETE', 'no')
-        self.replica_con.execute_command("GRAPH.CONFIG", "SET", 'ASYNC_DELETE', 'no')
-
-        # clear DB
-        self.conn.flushall()
-        self.source_con.execute_command("WAIT", "1", "0")
-
-    def tearDown(self):
-        # clear DB
-        self.source_con.flushall()
-        self.source_con.execute_command("WAIT", "1", "0")
-
-        assertStringPoolStats(self.source_con, 0, 0)
-        assertStringPoolStats(self.replica_con, 0, 0)
-
-    def query_and_wait(self, q, p=None):
-        if p is None:
-            p = {}
-
-        res = self.graph.query(q, p)
-
-        # the WAIT command forces master slave sync to complete
-        self.source_con.execute_command("WAIT", "1", "0")
-
-        return res
-
-    def test_intern_string_replication(self):
-        # both master and replica should be empty
-        assertStringPoolStats(self.source_con, 0, 0)
-        assertStringPoolStats(self.replica_con, 0, 0)
-
-        # replicate a node creation containing an intern string
-
-        s = LARGE_STRING
-
-        # create first node
-        p = {'s': s}
-        q = "CREATE ({value: intern($s)})"
-        self.query_and_wait(q, p)
-
-        assertStringPoolStats(self.source_con, 1, 1)
-        assertStringPoolStats(self.replica_con, 1, 1)
-
-        # replicate an addition of an intern string
-        q = "MATCH (n) SET n.s = intern($s)"
-        self.query_and_wait(q, p)
-
-        assertStringPoolStats(self.source_con, 1, 2)
-        assertStringPoolStats(self.replica_con, 1, 2)
-
-        # replicate deletion of an intern string
-        q = "MATCH (n) SET n.s = null"
-        self.query_and_wait(q)
-
-        assertStringPoolStats(self.source_con, 1, 1)
-        assertStringPoolStats(self.replica_con, 1, 1)
-
-        # replicate update of an intern string
-        q = "MATCH (n) SET n.value = intern('intern-string')"
-        self.query_and_wait(q)
-
-        assertStringPoolStats(self.source_con, 1, 1)
-        assertStringPoolStats(self.replica_con, 1, 1)
-
-        # replicate deletion if a node
-        q = "MATCH (n) DELETE n"
-        self.query_and_wait(q)
-
-        assertStringPoolStats(self.source_con, 0, 0)
-        assertStringPoolStats(self.replica_con, 0, 0)
